DebugClassLoaded is enable by default from 4.2, so it has to disabled on phpunit. I removed @runInSeparateProcess annotations on tests. ( It was a mistake, the real problem was that DebugClassLoaded enabled ) I added some documentation about PHP 8 Attributes on annotations.rst. I modified CHANGELOG.md I removed fos_http_cache.controller_resolver service and I use controller_resolver of Symfony instead I use PHP_VERSION_ID instead method_exists to check if we can use PHP Attributes. I renamed CatchAttibutesListener to Php8AttributesListener
