+++ phpunit.xml.dist
@@ -25,6 +25,19 @@
 
     <listeners>
         <listener class="\Mockery\Adapter\Phpunit\TestListener" />
+        <!--DebugClassLoaded is enable by default from 4.2
+            https://github.com/symfony/symfony/pull/28412
+            It has to be disabled to avoid tests failed
+        -->
+        <listener class="Symfony\Bridge\PhpUnit\SymfonyTestsListener">
+            <arguments>
+                <array>
+                    <element key="debug-class-loader">
+                        <integer>0</integer>
+                    </element>
+                </array>
+            </arguments>
+        </listener>
     </listeners>
 
     <php>
