--- src/Resources/config/catch_attributes.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" ?>
-
-<container xmlns="http://symfony.com/schema/dic/services"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-    <services>
-        <service id="fos_http_cache.controller_resolver" class="Symfony\Bundle\FrameworkBundle\Controller\ControllerResolver">
-            <argument type="service" id="service_container" />
-        </service>
-
-        <service id="fos_http_cache.event_listener.catch_attributes" class="FOS\HttpCacheBundle\EventListener\CatchAttributesListener">
-            <argument type="service" id="fos_http_cache.controller_resolver" />
-            <tag name="kernel.event_subscriber" />
-        </service>
-    </services>
-</container>
