+++ src/Resources/config/php8_attributes.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" ?>
+
+<container xmlns="http://symfony.com/schema/dic/services"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+    <services>
+        <service id="fos_http_cache.event_listener.php8_attributes" class="FOS\HttpCacheBundle\EventListener\Php8AttributesListener">
+            <argument type="service" id="controller_resolver" />
+            <tag name="kernel.event_subscriber" />
+        </service>
+    </services>
+</container>
