+++ tests/Functional/DependencyInjection/ServiceTest.php
@@ -42,9 +42,6 @@ class ServiceTest extends KernelTestCase
         return static::$kernel;
     }
 
-    /**
-     * @runInSeparateProcess
-     */
     public function testCanBeLoaded()
     {
         /** @var Container $container */
