+++ tests/Unit/DependencyInjection/Compiler/HashGeneratorPassTest.php
@@ -49,20 +49,10 @@ class HashGeneratorPassTest extends TestCase
         $config = $this->getBaseConfig();
         $this->extension->load([$config], $container);
         $this->userContextListenerPass->process($container);
-        if (method_exists(\ReflectionMethod::class, 'getAttributes')) {
-            if ($container->hasDefinition('service_container')) {
-                // symfony 3.3+
-                $this->assertCount(25, $container->getDefinitions());
-            } else {
-                $this->assertCount(26, $container->getDefinitions());
-            }
+        if (\PHP_VERSION_ID >= 80000) {
+            $this->assertCount(24, $container->getDefinitions());
         } else {
-            if ($container->hasDefinition('service_container')) {
-                // symfony 3.3+
-                $this->assertCount(23, $container->getDefinitions());
-            } else {
-                $this->assertCount(24, $container->getDefinitions());
-            }
+            $this->assertCount(23, $container->getDefinitions());
         }
     }
